# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unified Trading API",
    "description": "Consolidated API gateway \u2014 61 endpoints across 16 domains",
    "version": "0.1.0"
  },
  "paths": {
...
    "/market-data/candles": {
      "get": {
        "tags": [
          "market-data"
        ],
        "summary": "Get Candles",
        "description": "Get OHLCV candles for an instrument.",
        "operationId": "get_candles_market_data_candles_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue"
            }
          },
          {
            "name": "instrument",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1m",
              "title": "Timeframe"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Candles Market Data Candles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market-data/orderbook": {
      "get": {
        "tags": [
          "market-data"
        ],
        "summary": "Get Orderbook",
        "description": "Get order book snapshot.",
        "operationId": "get_orderbook_market_data_orderbook_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue"
            }
          },
          {
            "name": "instrument",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Orderbook Market Data Orderbook Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market-data/trades": {
      "get": {
        "tags": [
          "market-data"
        ],
        "summary": "Get Trades",
        "description": "Get recent trades.",
        "operationId": "get_trades_market_data_trades_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue"
            }
          },
          {
            "name": "instrument",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Trades Market Data Trades Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market-data/tickers": {
      "get": {
        "tags": [
          "market-data"
        ],
        "summary": "Get Tickers",
        "description": "Get all tickers for a venue.",
        "operationId": "get_tickers_market_data_tickers_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Tickers Market Data Tickers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/positions/active": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "Get Active Positions",
        "description": "Get currently active positions.",
        "operationId": "get_active_positions_positions_active_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Venue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Active Positions Positions Active Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/positions/summary": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "Get Position Summary",
        "description": "Get aggregated position summary across venues.",
        "operationId": "get_position_summary_positions_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Position Summary Positions Summary Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/positions/balances": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "Get Balances",
        "description": "Get account balances across venues.",
        "operationId": "get_balances_positions_balances_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Venue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Balances Positions Balances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/reporting/reports": {
      "get": {
        "tags": [
          "reporting"
        ],
        "summary": "Get Reports",
        "description": "Get generated reports.",
        "operationId": "get_reports_reporting_reports_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "report_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Report Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Reports Reporting Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reporting/settlements": {
      "get": {
        "tags": [
          "reporting"
        ],
        "summary": "Get Settlements",
        "description": "Get settlement reports.",
        "operationId": "get_settlements_reporting_settlements_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Settlements Reporting Settlements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reporting/reconciliation": {
      "get": {
        "tags": [
          "reporting"
        ],
        "summary": "Get Reconciliation",
        "description": "Get reconciliation results.",
        "operationId": "get_reconciliation_reporting_reconciliation_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Reconciliation Reporting Reconciliation Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}